backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
email-validator
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
httpx
openai
reportlab
Pillow
jinja2
aiofiles
sqlalchemy
alembic
psycopg2-binary
redis
celery
pytest
pytest-asyncio
requests

